refactor dropdown into its own sub-module
